tidy-up and beautifying
