--- src/routes/consent.ts
@@ -1,143 +1,0 @@
-import { NextFunction, Request, Response } from 'express'
-import {AdminApi as HydraAdminApi, AcceptConsentRequest, RejectRequest} from '@oryd/hydra-client'
-import url from 'url';
-import {logger} from '../config'
-
-const hydraAdminEndpoint = new HydraAdminApi(process.env.HYDRA_ADMIN_URL)
-
-export const getConsent = (
-    req: Request,
-    res: Response,
-    next: NextFunction
-  ) => {
-  // Parses the URL query
-  const query = url.parse(req.url, true).query;
-  // The challenge is used to fetch information about the consent request from ORY Hydra.
-  const challenge = query.consent_challenge;
-
-  hydraAdminEndpoint.getConsentRequest(String(challenge))
-    // This will be called if the HTTP request was successful
-    .then(function ({body}) {
-      // If a user has granted this application the requested scope, hydra will tell us to not show the UI.
-      if (body.skip) {
-        // You can apply logic here, for example grant another scope, or do whatever...
-
-        // Now it's time to grant the consent request. You could also deny the request if something went terribly wrong
-        const acceptConsentRequest = new AcceptConsentRequest()
-        // We can grant all scopes that have been requested - hydra already checked for us that no additional scopes
-        // are requested accidentally.
-        acceptConsentRequest.grantScope = body.requestedScope
-        // ORY Hydra checks if requested audiences are allowed by the client, so we can simply echo this.
-        acceptConsentRequest.grantAccessTokenAudience = body.requestedAccessTokenAudience
-        // The session allows us to set session data for id and access tokens
-        if ((body.requestedScope as any).indexOf('email') > -1) {
-          acceptConsentRequest.session = {
-            // This data will be available when introspecting the token. Try to avoid sensitive information here,
-            // unless you limit who can introspect tokens. (Therefore the scope-check above)
-            // access_token: { foo: 'bar' },
-
-            // This data will be available in the ID token.
-            // Most services need email-addresses, so let's include that.
-            idToken: { 
-              email: (body.context as any).addresses[0].value
-            },
-          }
-        }
-        return hydraAdminEndpoint.acceptConsentRequest(String(challenge), acceptConsentRequest).then(function (response:any) {
-          // All we need to do now is to redirect the user back to hydra!
-          res.redirect(response.body.redirectTo);
-        });
-      }
-
-      // If consent can't be skipped we MUST show the consent UI.
-      res.render('consent', {
-        csrfToken: req.csrfToken(),
-        challenge: challenge,
-        // We have a bunch of data available from the response, check out the API docs to find what these values mean
-        // and what additional data you have available.
-        requested_scope: body.requestedScope,
-        user: body.subject,
-        client: body.client,
-      });
-    })
-    // This will handle any error that happens when making HTTP calls to hydra
-    .catch(function (error:any) {
-      logger.error(error)
-      next(error);
-    });
-};
-
-export const postConsent = (
-    req: Request,
-    res: Response,
-    next: NextFunction
-  ) => {
-  // The challenge is now a hidden input field, so let's take it from the request body instead
-  var challenge = req.body.challenge;
-  // Let's see if the user decided to accept or reject the consent request..
-  if (req.body.submit != 'Allow access') {
-    // Looks like the consent request was denied by the user
-    const rejectConsentRequest = new RejectRequest()
-    rejectConsentRequest.error = 'access_denied'
-    rejectConsentRequest.errorDescription = 'The resource owner denied the request'
-    
-    return hydraAdminEndpoint.rejectConsentRequest( challenge, rejectConsentRequest)
-      .then(function (body:any) {
-        // All we need to do now is to redirect the browser back to hydra!
-        res.redirect(body.redirectTo);
-      })
-      // This will handle any error that happens when making HTTP calls to hydra
-      .catch(function (error:any) {
-        logger.error(error)
-        next(error);
-      }
-    );
-  }
-
-  var grant_scope = req.body.grant_scope
-  if (!Array.isArray(grant_scope)) {
-    grant_scope = [grant_scope]
-  }
-
-  // Seems like the user authenticated! Let's tell hydra...
-  hydraAdminEndpoint.getConsentRequest(challenge)
-  // This will be called if the HTTP request was successful
-    .then(({body}) => {
-      const acceptConsentRequest = new AcceptConsentRequest()
-      // We can grant all scopes that have been requested - hydra already checked for us that no additional scopes
-      // are requested accidentally.
-      acceptConsentRequest.grantScope = grant_scope
-      // ORY Hydra checks if requested audiences are allowed by the client, so we can simply echo this.
-      acceptConsentRequest.grantAccessTokenAudience = body.requestedAccessTokenAudience
-      // This tells hydra to remember this consent request and allow the same client to request the same
-      // scopes from the same user, without showing the UI, in the future.
-      acceptConsentRequest.remember = Boolean(req.body.remember),
-      // When this "remember" sesion expires, in seconds. Set this to 0 so it will never expire.
-      acceptConsentRequest.rememberFor = 3600
-      if (grant_scope.indexOf('email') > -1) {
-        // The session allows us to set session data for id and access tokens
-        acceptConsentRequest.session = {
-          // This data will be available when introspecting the token. Try to avoid sensitive information here,
-          // unless you limit who can introspect tokens.
-          // access_token: { foo: 'bar' },
-
-          // This data will be available in the ID token.
-          idToken: { 
-            email: (body.context as any).addresses[0].value
-          },
-        }
-      }
-
-
-      return hydraAdminEndpoint.acceptConsentRequest(challenge, acceptConsentRequest)
-        .then(function (response:any) {
-          // All we need to do now is to redirect the user back to hydra!
-          res.redirect(response.body.redirectTo);
-        })
-    })
-    // This will handle any error that happens when making HTTP calls to hydra
-    .catch((err) => {
-      logger.error(err)
-      next(err);
-    });
-};
