+++ views/login.hbs
@@ -6,7 +6,6 @@
         {{> messages messages=messages className="global"}}
         {{#if password}}
             <div id="login-password">
-                <div style="margin-bottom: 8px">Login</div>
                 {{> form submitLabel="Sign in" form=password}}
             </div>
         {{/if}}
