+++ views/registration.hbs
@@ -7,7 +7,6 @@
         {{> messages messages=messages className="global"}}
         {{#if password}}
             <div id="registration-password">
-                <div style="margin-bottom: 8px">Register</div>
                 {{> form submitLabel="Sign up" form=password}}
             </div>
         {{/if}}
